Fixed warnings in debug code.
git-svn-id: file:///svn/nsd/trunk@1222 a26ef69c-88ff-0310-839f-98b793d9c207

+++ query.c
@@ -353,7 +353,7 @@ process_edns (struct query *q, uint8_t *qptr)
 				ARCOUNT(q) = 0;
 
 				DEBUG(DEBUG_QUERY, 2,
-				      (stderr, "EDNS0 maxlen = %u\n", q->maxlen));
+				      (stderr, "EDNS0 maxlen = %lu\n", (unsigned long) q->maxlen));
 			}
 		}
 	}
